Fixed width

+++ users.php
@@ -4,6 +4,7 @@ session_get_cookie_params();
 if (!isset($_SESSION['user'])) {
     header('Location: ./index.php');
 }
+$current_user = $_SESSION["user"];
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -59,9 +60,15 @@ if (!isset($_SESSION['user'])) {
                         <a href="./users/edit.php?id=<?= $user["id"] ?>"><svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-8 h-8 stroke-emerald-600">
                                 <path stroke-linecap="round" stroke-linejoin="round" d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L10.582 16.07a4.5 4.5 0 01-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 011.13-1.897l8.932-8.931zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0115.75 21H5.25A2.25 2.25 0 013 18.75V8.25A2.25 2.25 0 015.25 6H10" />
                             </svg></a>
-                        <a href="./users/delete.php?id=<?= $user["id"] ?>"><svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-8 h-8 stroke-red-600">
-                                <path stroke-linecap="round" stroke-linejoin="round" d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0" />
-                            </svg></a>
+                        <?php if ($user["email"] != $current_user) : ?>
+                            <a href="./users/delete.php?id=<?= $user["id"] ?>"><svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-8 h-8 stroke-red-600">
+                                    <path stroke-linecap="round" stroke-linejoin="round" d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0" />
+                                </svg></a>
+                        <?php else : ?>
+                            <button id="user-delete" class="cursor-not-allowed"><svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-8 h-8 stroke-red-600">
+                                    <path stroke-linecap="round" stroke-linejoin="round" d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0" />
+                                </svg></button>
+                        <?php endif; ?>
 
                     </div>
                 </div>
@@ -75,19 +82,21 @@ if (!isset($_SESSION['user'])) {
                     <path stroke-linecap="round" stroke-linejoin="round" d="M9 12.75L11.25 15 15 9.75M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
                 </svg>
                 <?php if (isset($_GET['action'])) : ?>
-                    <span class="text-white font-semibold"><?php
-                                                            switch ($_GET['action']) {
-                                                                case 'new':
-                                                                    echo "Usuario creado correctamente";
-                                                                    break;
-                                                                case 'edit':
-                                                                    echo "Usuario editado correctamente";
-                                                                    break;
-                                                                case 'delete':
-                                                                    echo "Usuario eliminado correctamente";
-                                                                    break;
-                                                            }
-                                                            ?></span>
+                    <span class="text-white font-semibold">
+                        <?php
+                        switch ($_GET['action']) {
+                            case 'new':
+                                echo "Usuario creado correctamente";
+                                break;
+                            case 'edit':
+                                echo "Usuario editado correctamente";
+                                break;
+                            case 'delete':
+                                echo "Usuario eliminado correctamente";
+                                break;
+                        }
+                        ?>
+                    </span>
                 <?php endif; ?>
             </div>
             <div class="h-1/5 rounded-b-xl">
@@ -102,12 +111,21 @@ if (!isset($_SESSION['user'])) {
                     <path stroke-linecap="round" stroke-linejoin="round" d="M9.75 9.75l4.5 4.5m0-4.5l-4.5 4.5M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
                 </svg>
 
-                    <span class="text-white font-semibold">Se ha producido un error</span>
+                <span class="text-white font-semibold">Se ha producido un error</span>
             </div>
             <div class="h-1/5 rounded-b-xl">
                 <div class="w-full h-full rounded-b-xl bg-red-800 loader"></div>
             </div>
         <?php endif; ?>
+        <script>
+            function userWarning() {
+                alert("No puedes eliminar tu propio usuario");
+            }
+            const userDelete = document.getElementById("user-delete");
+            if (userDelete) {
+                userDelete.addEventListener("click", userWarning);
+            }
+        </script>
 </body>
 
 </html>
